% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{Add_cv_folds}
\alias{Add_cv_folds}
\title{Add cross-validation fold assignments to a data.table}
\usage{
Add_cv_folds(dt, k = 5, seed = 123)
}
\arguments{
\item{dt}{data.table to which folds should be added.}

\item{k}{Integer, number of folds (default = 5).}

\item{seed}{Integer, random seed for reproducibility (default = 123).}
}
\value{
The input data.table \code{dt} with an added column \code{fold}.
}
\description{
Randomly assigns each row of a data.table to one of \code{k} cross-validation folds.
}
