% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{assign_greedy}
\alias{assign_greedy}
\title{A helper function for egg assignment based on prior probabilities}
\usage{
assign_greedy(dt, nestleftover = NULL, pen_meta)
}
\arguments{
\item{dt}{Prior data.table \verb{[eid, ani, p]}.}

\item{nestleftover}{data.table of leftover priors \verb{[eid, ani, p]}.}

\item{pen_meta}{data.table with animal metadata.}
}
\value{
data.table of selected candidate-egg pairs.
}
\description{
First assign nest eggs by searching for the globally highest probabilities,
then the p of unselected candidates of assigned eggs (leftover) will be add on to their p for other eggs for competition
floor eggs assignment can include leftover from nest eggs assignments (but not the other way around).
}
