% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_date}
\alias{check_date}
\title{Validate date is in yyyy-mm-dd format}
\usage{
check_date(input)
}
\arguments{
\item{input}{Character string or Date}
}
\value{
Invisible TRUE if the input passes the check.
}
\description{
Checks if input is a valid date in the format YYYY-MM-DD.
#'
}
