% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_datetime}
\alias{check_datetime}
\title{Validate POSIXct datetime input}
\usage{
check_datetime(input)
}
\arguments{
\item{input}{A single R object. Expected to be POSIXct datetime.}
}
\value{
Invisible TRUE if the input passes the check.
}
\description{
This helper function checks whether the provided input is a POSIXct datetime object.
}
