% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{check_time}
\alias{check_time}
\title{Validate time is in hh:mm:ss format}
\usage{
check_time(input)
}
\arguments{
\item{input}{Character string or hms}
}
\value{
Invisible TRUE if the input passes the check.
}
\description{
Checks if input is a valid time in the format hh:mm:ss.
#'
}
