% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{CV_pen}
\alias{CV_pen}
\title{Cross-validation for pen-level trusted egg-animal pairs}
\usage{
CV_pen(
  pen_trusted_dat,
  reps = 2,
  k = 5,
  seed = 123,
  p,
  negg,
  meta,
  ani_info,
  index_floornest = 99,
  ani_impeggs,
  weight_imd = 1,
  factor_imd = 1,
  p_perc_dups = 50,
  nearby_days = 2,
  penalty_weight = 0.3,
  max_penalty = 0.6,
  weight_nl = 0.8,
  factor_nl = 1,
  weight_np = 0.2,
  factor_np = 1,
  norm_result = "prob",
  flat = FALSE
)
}
\arguments{
\item{pen_trusted_dat}{Data table of trusted egg-animal pairs with at least columns: \code{eid}, \code{ani}.}

\item{reps}{Number of CV repetitions (default = 2).}

\item{k}{Number of folds per repetition (default = 5).}

\item{p}{Pen ID.}

\item{negg}{data.table of egg counts per nest/floor per date.}

\item{meta}{data.table of autonest metadata.}

\item{ani_info}{data.table with animal presence/absence info.}

\item{index_floornest}{Integer ID used for floor eggs (default: 99).}

\item{ani_impeggs}{data.table of imputed laying dates.}

\item{weight_imd}{For prior update using imputed laying dates: Weight (default: 1).}

\item{factor_imd}{For prior update using imputed laying dates: factor (default: 1).}

\item{p_perc_dups}{Probability percentage for duplicate records (default: 50).}

\item{nearby_days}{Integer, for nearby lay pattern update: Number of days before & after the date to include (default: 2).}

\item{penalty_weight}{For nearby lay pattern update: Numeric penalty for switching nests (default: 0.3).}

\item{max_penalty}{For nearby lay pattern update: Maximum nest-switching penalty allowed (default: 0.6).}

\item{weight_nl}{For nearby lay pattern update: Weight (default: 0.8).}

\item{factor_nl}{For nearby lay pattern update: Factor (default: 1).}

\item{weight_np}{For nest preference update: Weight (default: 0.2).}

\item{factor_np}{For nest preference update: Factor (default: 1).}

\item{norm_result}{Character, either \code{"prob"} to normalize probabilities (default) or \code{"assign"}.}

\item{flat}{TRUE: to get only flat priors; FALSE: to get Bayesian updated probabilities (default: FALSE).}
}
\value{
Data table of masked trusted egg-animal pairs with evaluation results.
}
\description{
Performs cross-validation by masking trusted egg-animal pairs in a pen
and evaluating the prior assignment procedure.
}
