% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_CV}
\alias{get_CV}
\title{Get CV results from function "CV_pen"}
\usage{
get_CV(res_list)
}
\arguments{
\item{res_list}{A list contains both or either "prob" and "assign" output}
}
\value{
A data.table of results, \code{pen}, \code{TF}(True/False), \code{prop}(proportion), \code{type}(prob/assign)
}
\description{
Get CV results from function "CV_pen"
}
