% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_ENdt}
\alias{get_ENdt}
\title{Get prob/assign results as a data table}
\usage{
get_ENdt(res_list)
}
\arguments{
\item{res_list}{A list contains "pen_priors" and "pen_trusted" output by pen}
}
\value{
A data table of results,\code{eiddate}(date of egg collection), \code{ani},
\code{prior}(for trusted & results type = "assign", prior = 1; results type = "prob", prior is the posterior probability),
\code{layingtime}(fro
}
\description{
Get prob/assign results as a data table
}
