% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_ENmatrix}
\alias{get_ENmatrix}
\title{Get prob/assign results as a \verb{[date x ani]} matrix}
\usage{
get_ENmatrix(res_list)
}
\arguments{
\item{res_list}{A list contains "pen_priors" and "pen_trusted" output by pen}
}
\value{
A matrix of results \verb{[date x ani]}
}
\description{
Get prob/assign results as a \verb{[date x ani]} matrix
}
