% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preps.R
\name{get_good_hand_eggcount}
\alias{get_good_hand_eggcount}
\title{Combine hand-counted and automated egg data for validation}
\usage{
get_good_hand_eggcount(
  meta,
  hand,
  from = NULL,
  to = NULL,
  hand_nest_colnames = c("Nest1", "Nest2", "Nest3", "Nest4"),
  Nestnumbers = 1:64,
  num_per_pen = 4,
  timezone = "Europe/Berlin",
  collect_min = 20,
  fakeegg = FALSE
)
}
\arguments{
\item{meta}{A \code{data.table} containing raw autonest data.}

\item{hand}{\code{data.table} containing hand-count egg data.}

\item{from, to}{\code{Date} or \code{character(1)}, optional date range.}

\item{Nestnumbers}{Vector of nest numbers (default: \code{1:64}).}

\item{num_per_pen}{Number of nests per pen (default: 4).}

\item{timezone}{Time zone used for datetime conversion (default: \code{"Europe/Berlin"}).}

\item{collect_min}{Total time needed in minutes for egg collection (default: 20).}
}
\value{
A \code{data.table} containing hand and automatic egg counts per nest and time interval.
with the following key columns:
\itemize{
\item \code{Date}: Date of observation
\item \code{Pen}: pen ID
\item \code{Nest}: nest number
\item \code{Nhand}: manually counted eggs
\item \code{Nfloor}: manually counted floor eggs
\item \code{dtm_a}: algorithm adjusted autonest time (from fakeegg transponder if available)
\item \code{dtm_h} : hand recorded datetime, replaced by reliable dtm_a (if available)
\item \code{dtm_hpre}, : previous record of dtm_h
\item \code{Nauto_all}: all autonest eggs within time interval
\item \code{Nauto_filt}: filtered autonest eggs (no flag -> F_combined = FALSE)
}
}
\description{
Reads and integrates manually counted egg data with automatic
nest system data to generate a harmonized table for evaluating egg count accuracy.
}
