% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{get_imputed_laydates}
\alias{get_imputed_laydates}
\title{Impute missing laying dates for a given animal}
\usage{
get_imputed_laydates(
  pen_meta,
  aniid,
  max_gap_hour = 200,
  nstate = 4,
  max_iter = 10,
  stop_pegg = 8,
  min_pegg = 1.5,
  thrd_laydiffh = 24,
  plot = FALSE
)
}
\arguments{
\item{pen_meta}{data.table with at least the columns: \code{ani}, \code{Date}, \code{Eggsignal},
\code{F_combined}, \code{datelay}, \code{laydiffh_pre_all}, \code{laydiffh_nxt_all}, \code{Layingtime}}

\item{aniid}{Character or numeric ID of the animal to process}

\item{max_gap_hour}{Maximum allowed gap in hours before considering a "break" (default 200)}

\item{nstate}{Number of states for HMM model (default 4)}

\item{max_iter}{Maximum iterations for HMM fitting (default 10)}

\item{stop_pegg}{Stopping criteria for fitting HMM when the prob of laying reached a certain percentage (default 8)}

\item{min_pegg}{Minimum prob of laying (percentage) for a state to be considered as laying (default 1.5)}

\item{thrd_laydiffh}{Minimum hours threshold for laydiff to consider imputation (default 24)}

\item{plot}{TRUE or FALSE to plot the output (default FALSE)}
}
\value{
data.table of imputed laying dates and time (\code{Date}, \code{datelay}, \code{Layingtime}) and \code{ani} ID,
\code{datelay} is POSIXct datetime, combined by imputed laying date and mean laying time
including \code{imp} as information and \code{p_perc} as the prob of laying (percentage). Returns \code{NULL} if insufficient data or fit fails.
}
\description{
Imputes likely laying dates for a single animal based on its
observed eggs in \code{pen_meta}, using Hidden Markov Model fitting.
}
