% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_matrix}
\alias{get_matrix}
\title{Get prob/assign results as a \verb{[date x ani]} matrix}
\usage{
get_matrix(res_list, type = "assign")
}
\arguments{
\item{res_list}{A list contains both or either "prob" and "assign" output}
}
\value{
A matrix of results \verb{[date x ani]}
}
\description{
Get prob/assign results as a \verb{[date x ani]} matrix
}
