% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_prob}
\alias{get_prob}
\title{Get prob/assign results per egg from function "process_pen"}
\usage{
get_prob(res_list, type = "prob")
}
\arguments{
\item{res_list}{A list contains both or either "prob" and "assign" output}
}
\value{
A data.table of results, \code{eid}(egg id), \code{ani}, \code{prior}, \code{pen}, \code{Date}, \code{type}(prob/assign)
}
\description{
Get prob/assign results per egg from function "process_pen"
}
