% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_rmegg}
\alias{get_rmegg}
\title{Get removed eggs from function "process_pen"}
\usage{
get_rmegg(res_list)
}
\arguments{
\item{res_list}{A list contains "pen_priors" and "pen_trusted" output by pen}
}
\value{
A data.table of results,\code{Nest}, \code{eid}(egg id), \code{date}, \code{pen}
}
\description{
Get removed eggs from function "process_pen"
}
