% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_sumprob}
\alias{get_sumprob}
\title{Get summed prob & assign results over a period from function "process_pen"}
\usage{
get_sumprob(res_list)
}
\arguments{
\item{res_list}{A list contains both or either "prob" and "assign" output}
}
\value{
A data.table of results, \code{ani}, \code{sum_prior}(sum of prob), \code{type}(prob/assign)
}
\description{
Get summed prob & assign results over a period from function "process_pen"
}
