% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_trusted}
\alias{get_trusted}
\title{Get trusted set from function "process_pen"}
\usage{
get_trusted(res_list)
}
\arguments{
\item{res_list}{A list contains both or either "prob" and "assign" output}
}
\value{
A data.table of results,\code{eid}(egg id), \code{ani}, \code{layingtime}(from autonest)
\code{type}(primary/secondary), \code{date}, \code{pen}, \code{fold}(if CV=TRUE)
}
\description{
Get trusted set from function "process_pen"
}
