% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preps.R
\name{get_trusted_autonest}
\alias{get_trusted_autonest}
\title{Identify trusted egg-animal pairs from autonest data}
\usage{
get_trusted_autonest(eggs, pen_meta, from, to)
}
\arguments{
\item{eggs}{\code{data.table} with columns: \code{Nest}, \code{eggid}.}

\item{pen_meta}{\code{data.table} with autonest metadata.}

\item{from}{POSIXct start datetime.}

\item{to}{POSIXct end datetime.}
}
\value{
\code{data.table} with columns: \code{eid}, \code{ani}, \code{layingtime}, \code{type}.
}
\description{
Filters autonest data to identify "trusted" candidate animals for each egg
within a given datetime range.
}
