% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{load_results}
\alias{load_results}
\title{Load results files RData by type}
\usage{
load_results(
  types = c("prob", "assign"),
  pattern = "\\\\.RData$",
  directory = ".",
  new = 1
)
}
\arguments{
\item{types}{Character of types to load (e.g., c("prob", "assign") or one of them)}

\item{pattern}{Pattern to match files (default: looks for .RData files)}

\item{directory}{Directory to search (default: current working directory)}

\item{new}{The rank position of the newest files sorted by date (default: 1)}
}
\value{
A list with elements for each type containing loaded objects
}
\description{
Load results files RData by type
}
