% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{make_naive_prior}
\alias{make_naive_prior}
\title{Create a naive prior probability matrix}
\usage{
make_naive_prior(eggid, cand_ani)
}
\arguments{
\item{eggid}{Vector of egg IDs}

\item{cand_ani}{Vector of candidate animal IDs}
}
\value{
Matrix of size \verb{[length(eggid) x length(cand_ani)]} with equal probabilities
}
\description{
Generates a matrix of equal probabilities for all candidate animals
for each egg. Each row corresponds to an egg, each column to a candidate animal.
}
