% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{modify_eid}
\alias{modify_eid}
\title{modify eid for floor eggs to get get unique eid even across pens}
\usage{
modify_eid(respp)
}
\arguments{
\item{dat}{A data.table contains \code{eid}(egg id), \code{pen}}
}
\description{
modify eid for floor eggs to get get unique eid even across pens
}
