% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{Norm_prior}
\alias{Norm_prior}
\title{Normalise prior probabilities}
\usage{
Norm_prior(prior, result = "prob", index_floornest, pen_meta)
}
\arguments{
\item{prior}{Prior matrix \verb{[egg x candidate]}.}

\item{result}{Character, either \code{"prob"} to normalize probabilities (default) or \code{"assign"} to assign top candidate per egg.}

\item{index_floornest}{Numeric index for floor eggs.}

\item{pen_meta}{data.table with animal metadata.}
}
\value{
Prior matrix with row sums = 1 (\code{result = "prob"}) or with top candidate assigned (\code{result = "assign"}).
}
\description{
Ensures that the prior matrix rows and cols sum to 1 and optionally converts it to an
assignment matrix where each egg is assigned to the most probable candidate.
}
