% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preps.R
\name{prep_data}
\alias{prep_data}
\title{Prepare autonest laying data}
\usage{
prep_data(
  meta_ori,
  ot_min = NULL,
  dateformat = "\%d.\%m.\%Y",
  from = NULL,
  to = NULL
)
}
\arguments{
\item{ot_min}{How many minutes to correct for the autonest clock (default NULL).}

\item{dateformat}{Date format in meta (default "\%d.\%m.\%Y").}

\item{from, to}{\code{Date} or \code{character(1)}, optional date range (format yyyy-mm-dd.}

\item{meta}{A \code{data.table} containing raw autonest data.}
}
\value{
A \code{data.table} with cleaned time/date columns (\code{hms}/\code{Date}), and
additional columns: \code{ani} and \code{type}.
}
\description{
Cleans and formats raw autonest metadata, converts times to \code{hms} objects,
converts dates to \code{Date} objects, and marks primary vs secondary animals.
}
