% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{update_prior_nestlay}
\alias{update_prior_nestlay}
\title{Update prior matrix using laying patterns}
\usage{
update_prior_nestlay(
  prior,
  eggs,
  cand_ani,
  pen_meta,
  from,
  to,
  index_floornest,
  nearby_days = 2,
  penalty_weight = 0.3,
  max_penalty = 0.6,
  weight = 0.8,
  factor = 1
)
}
\arguments{
\item{prior}{Prior matrix \verb{[egg x candidate]}.}

\item{eggs}{data.table with at least: \code{Nest}, \code{eggid}.}

\item{cand_ani}{Vector of candidate animal IDs.}

\item{pen_meta}{data.table with at least: \code{ani}, \code{Nestnumber}, \code{Date}, \code{Eggsignal}, \code{F_combined}, \code{laydiffh_pre_all}.}

\item{from}{POSIXct start datetime.}

\item{to}{POSIXct end datetime.}

\item{index_floornest}{Numeric pen index for floor eggs.}

\item{nearby_days}{Integer, number of days before & after \code{date} to include (default 2).}

\item{penalty_weight}{Numeric penalty for switching nests (default 0.3).}

\item{max_penalty}{Maximum nest-switching penalty allowed (default 0.6).}

\item{weight}{Numeric weight to apply to laying patterns of nearby days (default 0.8).}

\item{factor}{Numeric factor to scale the prior update (default 1).}
}
\value{
Updated prior matrix.
}
\description{
Adjusts prior probabilities based on candidates' laying patterns
observed in nearby days from autonest data.
}
