% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preps.R
\name{prep_data}
\alias{prep_data}
\title{Prepare autonest laying data}
\usage{
prep_data(meta, ot_min = NULL)
}
\arguments{
\item{meta}{A \code{data.table} containing raw autonest data.}

\item{ot_min}{How many minutes to correct for the autonest clock (default NULL).}
}
\value{
A \code{data.table} with cleaned time/date columns (\code{hms}/\code{Date}), and
additional columns: \code{ani} and \code{type}.
}
\description{
Cleans and formats raw autonest metadata, converts times to \code{hms} objects,
converts dates to \code{Date} objects, and marks primary vs secondary animals.
}
